# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "contact": {
      "email": "eti@cisco.io",
      "name": "API Support"
    },
    "description": "This is a Kosha REST service for exposing many teamwork features as REST APIs with better consistency, observability etc",
    "termsOfService": "http://swagger.io/terms/",
    "title": "Teamwork Connector API",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
...
    "/api/v1/projects/{project_id}/timeentry": {
      "get": {
        "parameters": [
          {
            "description": "Enter project id",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/models.ReturnedTimeEntries"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "timeentry"
        ],
        "description": "List all time entries in the project\nPlease refer to https://apidocs.teamwork.com/docs/teamwork/aa65992887407-retrieve-all-time-entries-for-a-project for more parameter options.",
        "summary": "Get all time entries associated with specified project"
      },
      "post": {
        "parameters": [
          {
            "description": "Enter project id",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "timeentry"
        ],
        "description": "Create a time entry for a project in the system\nPlease refer to https://apidocs.teamwork.com/docs/teamwork/0a02f6155b979-create-a-time-entry for more parameter options.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/models.CreateTimeEntry"
              }
            }
          },
          "description": "Enter time entry properties"
        },
        "summary": "Create a time entry for a project"
      }
    },
...
    "/api/v1/timeentries/{id}": {
      "delete": {
        "parameters": [
          {
            "description": "Enter time entry id",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "timeentry"
        ],
        "description": "Remove a time entry\nPlease refer to https://apidocs.teamwork.com/docs/teamwork/c16cf6a573f70-delete-a-time-entry for more parameter options.",
        "summary": "Delete a time entry"
      },
      "put": {
        "parameters": [
          {
            "description": "Enter time entry id",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/models.CreatedTimeEntry"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "timeentry"
        ],
        "description": "Add an update to a time entry\nPlease refer to https://apidocs.teamwork.com/docs/teamwork/e1320843e680e-update-a-time-entry for more parameter options.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/models.CreateTimeEntry"
              }
            }
          },
          "description": "Enter time entry properties"
        },
        "summary": "Update a time entry"
      }
    }
  },
  "components": {
...
    "schemas": {
...
      "models.CreateTimeEntry": {
        "properties": {
          "time-entry": {
            "$ref": "#/components/schemas/models.TimeEntry"
          }
        },
        "type": "object"
      },
      "models.CreatedTimeEntry": {
        "properties": {
          "status": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "models.ReturnedTimeEntries": {
        "properties": {
          "STATUS": {
            "type": "string"
          },
          "time-entries": {
            "items": {
              "$ref": "#/components/schemas/models.ReturnedTimeEntryObject"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
      "models.ReturnedTimeEntryObject": {
        "properties": {
          "tags": {},
          "company-id": {
            "description": "TicketId string `json:\"ticket-id,omitempty\"`\nCreatedAt string `json:\"createdAt,omitempty\"`\nTaskIsPrivate string `json:\"taskIsPrivate,omitempty\"`\nParentTaskId string `json:\"parentTaskId,omitempty\"`",
            "type": "string"
          },
          "company-name": {
            "description": "CanEdit bool `json:\"canEdit,omitempty\"`\nTaskEstimatedTime string `json:\"taskEstimatedTime,omitempty\"`",
            "type": "string"
          },
          "date": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "has-start-time": {
            "type": "string"
          },
          "hours": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "invoiceNo": {
            "type": "string"
          },
          "isbillable": {
            "type": "string"
          },
          "minutes": {
            "description": "ParentTaskName string `json:\"parentTaskName,omitempty\"`\nDateUserPerspective string `json:\"dateUserPerspective,omitempty\"`",
            "type": "string"
          },
          "person-first-name": {
            "type": "string"
          },
          "person-id": {
            "type": "string"
          },
          "person-last-name": {
            "type": "string"
          },
          "project-id": {
            "type": "string"
          },
          "project-name": {
            "type": "string"
          },
          "project-status": {
            "type": "string"
          },
          "todo-item-id": {
            "description": "TaskTags interface{} `json:\"task-tags,omitempty\"`\nTaskIsSubTask string `json:\"taskIsSubTask,omitempty\"`",
            "type": "string"
          },
          "todo-item-name": {
            "type": "string"
          },
          "todo-list-id": {
            "description": "IsBilled string `json:\"isbilled,omitempty\"`\nUpdatedDate string `json:\"updated-date,omitempty\"`",
            "type": "string"
          },
          "todo-list-name": {
            "description": "TaskListId string `json:\"tasklistId,omitempty\"`",
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "models.TimeEntry": {
        "properties": {
          "tags": {
            "type": "string"
          },
          "date": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "hours": {
            "type": "integer"
          },
          "isbillable": {
            "type": "boolean"
          },
          "minutes": {
            "type": "integer"
          },
          "person-id": {
            "type": "integer"
          },
          "time": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
    }
  }
}